+++ OneBlink.SDK/models/FormSubmissionEvent.cs
@@ -71,7 +71,8 @@ namespace OneBlink.SDK.Model
             bool requiresAllConditionallyExecutePredicates = false,
             bool groupFiles = false,
             List<string> excludedElementIds = default(List<string>),
-            bool? usePagesAsBreaks = null)
+            bool? usePagesAsBreaks = null,
+            bool preventExtensionlessAttachments = false)
         {
             FormSubmissionEventConfigration fseconfig = new FormSubmissionEventConfigration();
             fseconfig.environmentId = environmentId;
@@ -83,6 +84,7 @@ namespace OneBlink.SDK.Model
             fseconfig.groupFiles = groupFiles;
             fseconfig.excludedElementIds = excludedElementIds;
             fseconfig.usePagesAsBreaks = true;
+            fseconfig.preventExtensionlessAttachments = preventExtensionlessAttachments;
             FormSubmissionEvent trim = new FormSubmissionEvent();
             trim.type = "TRIM";
             trim.isDraft = isDraft;